k210_pac
Module aes

AES Accelerator

Modules

aad_data

Additional authenticated data

aad_num

GCM additional authenticated data count in bytes, minus one

data_in_flag

Data can input flag

data_out_flag

Data can output flag

dma_sel

DMA select

en

AES module enable

encrypt_sel

Encryption or decryption select

endian

Endian control

finish

Finished status

gcm_in_tag

GCM input tag for comparison with the calculated tag

gcm_out_tag

Computed GCM output tag

iv

Initialisation Vector (96 bit for GCM, 128 bit for CBC)

key

1st-4th word of key

key_ext

5th-8th word of key

mode_ctl

AES mode register

out_data

Plaintext/ciphertext output data

pc_num

Plaintext/ciphertext input data count in bytes, minus one

tag_chk

Tag check status

tag_clear

Tag clear (a write to this register clears the tag_chk status)

tag_in_flag

Can input tag (when using GCM)

text_data

Plaintext/ciphertext input data

Structs

RegisterBlock

Register block

Type Aliases

AAD_DATA

Additional authenticated data

AAD_NUM

GCM additional authenticated data count in bytes, minus one

DATA_IN_FLAG

Data can input flag

DATA_OUT_FLAG

Data can output flag

DMA_SEL

DMA select

EN

AES module enable

ENCRYPT_SEL

Encryption or decryption select

ENDIAN

Endian control

FINISH

Finished status

GCM_IN_TAG

GCM input tag for comparison with the calculated tag

GCM_OUT_TAG

Computed GCM output tag

IV

Initialisation Vector (96 bit for GCM, 128 bit for CBC)

KEY

1st-4th word of key

KEY_EXT

5th-8th word of key

MODE_CTL

AES mode register

OUT_DATA

Plaintext/ciphertext output data

PC_NUM

Plaintext/ciphertext input data count in bytes, minus one

TAG_CHK

Tag check status

TAG_CLEAR

Tag clear (a write to this register clears the tag_chk status)

TAG_IN_FLAG

Can input tag (when using GCM)

TEXT_DATA

Plaintext/ciphertext input data