k210_pac::aes::data_out_flag
Type Alias DATA_OUT_FLAG_R

pub type DATA_OUT_FLAG_R = R<bool, DATA_OUT_FLAG_A>;

Reader of field data_out_flag

Aliased Type

struct DATA_OUT_FLAG_R { /* private fields */ }

Implementations

impl DATA_OUT_FLAG_R

pub fn variant(&self) -> DATA_OUT_FLAG_A

Get enumerated values variant

pub fn is_cannot_output(&self) -> bool

Checks if the value of the field is CANNOT_OUTPUT

pub fn is_can_output(&self) -> bool

Checks if the value of the field is CAN_OUTPUT