k210_pac::aes::mode_ctl
Type Alias CIPHER_MODE_R

pub type CIPHER_MODE_R = R<u8, CIPHER_MODE_A>;

Reader of field cipher_mode

Aliased Type

struct CIPHER_MODE_R { /* private fields */ }

Implementations

impl CIPHER_MODE_R

pub fn variant(&self) -> Variant<u8, CIPHER_MODE_A>

Get enumerated values variant

pub fn is_ecb(&self) -> bool

Checks if the value of the field is ECB

pub fn is_cbc(&self) -> bool

Checks if the value of the field is CBC

pub fn is_gcm(&self) -> bool

Checks if the value of the field is GCM