Type Definition k210_pac::apu::ctl::R

pub type R = R<u32, CTL>;

Reader of register ctl

Implementations

impl R

pub fn dir_search_en(&self) -> DIR_SEARCH_EN_R

Bit 0 - Sound direction searching enable bit

pub fn search_path_reset(&self) -> SEARCH_PATH_RESET_R

Bit 1 - Reset all control logic on direction search processing path

pub fn stream_gen_en(&self) -> STREAM_GEN_EN_R

Bit 4 - Valid voice sample stream generation enable bit

pub fn voice_gen_path_reset(&self) -> VOICE_GEN_PATH_RESET_R

Bit 5 - Reset all control logic on voice stream generating path

pub fn update_voice_dir(&self) -> UPDATE_VOICE_DIR_R

Bit 6 - Switch to a new voice source direction