k210_pac::apu::ctl
Type Alias W

pub type W = W<u32, CTL>;

Writer for register ctl

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn dir_search_en(&mut self) -> DIR_SEARCH_EN_W<'_>

Bit 0 - Sound direction searching enable bit

pub fn search_path_reset(&mut self) -> SEARCH_PATH_RESET_W<'_>

Bit 1 - Reset all control logic on direction search processing path

pub fn stream_gen_en(&mut self) -> STREAM_GEN_EN_W<'_>

Bit 4 - Valid voice sample stream generation enable bit

pub fn voice_gen_path_reset(&mut self) -> VOICE_GEN_PATH_RESET_W<'_>

Bit 5 - Reset all control logic on voice stream generating path

pub fn update_voice_dir(&mut self) -> UPDATE_VOICE_DIR_W<'_>

Bit 6 - Switch to a new voice source direction

pub fn we_dir_search_en(&mut self) -> WE_DIR_SEARCH_EN_W<'_>

Bit 8 - Write enable for we_dir_search_en parameter

pub fn we_search_path_rst(&mut self) -> WE_SEARCH_PATH_RST_W<'_>

Bit 9 - Write enable for we_search_path_rst parameter

pub fn we_stream_gen(&mut self) -> WE_STREAM_GEN_W<'_>

Bit 10 - Write enable for we_stream_gen parameter

pub fn we_voice_gen_path_rst(&mut self) -> WE_VOICE_GEN_PATH_RST_W<'_>

Bit 11 - Write enable for we_voice_gen_path_rst parameter

pub fn we_update_voice_dir(&mut self) -> WE_UPDATE_VOICE_DIR_W<'_>

Bit 12 - Write enable for we_update_voice_dir parameter