k210_pac::dmac::channel::cfg
Type Alias LOCK_CH_L_R

pub type LOCK_CH_L_R = R<u8, LOCK_CH_L_A>;

Reader of field lock_ch_l

Aliased Type

struct LOCK_CH_L_R { /* private fields */ }

Implementations

impl LOCK_CH_L_R

pub fn variant(&self) -> Variant<u8, LOCK_CH_L_A>

Get enumerated values variant

pub fn is_dma_transfer(&self) -> bool

Checks if the value of the field is DMA_TRANSFER

pub fn is_block_transfer(&self) -> bool

Checks if the value of the field is BLOCK_TRANSFER

pub fn is_transaction(&self) -> bool

Checks if the value of the field is TRANSACTION