k210_pac::dmac::channel::cfg
Type Alias R

pub type R = R<u64, CFG>;

Reader of register cfg

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn src_multblk_type(&self) -> SRC_MULTBLK_TYPE_R

Bits 0:1 - Source multi-block transfer type

pub fn dst_multblk_type(&self) -> DST_MULTBLK_TYPE_R

Bits 2:3 - Destination multi-block transfer type

pub fn tt_fc(&self) -> TT_FC_R

Bits 32:34 - Transfer type and flow control

pub fn hs_sel_src(&self) -> HS_SEL_SRC_R

Bit 35 - Source software or hardware handshaking select

pub fn hs_sel_dst(&self) -> HS_SEL_DST_R

Bit 36 - Destination software or hardware handshaking select

pub fn src_hwhs_pol(&self) -> SRC_HWHS_POL_R

Bit 37 - Source hardware handshaking interface polarity

pub fn dst_hwhs_pol(&self) -> DST_HWHS_POL_R

Bit 38 - Destination hardware handshaking interface polarity

pub fn src_per(&self) -> SRC_PER_R

Bits 39:42 - Assign a hardware handshaking interface to source of channel

pub fn dst_per(&self) -> DST_PER_R

Bits 44:47 - Assign a hardware handshaking interface to destination of channel

pub fn ch_prior(&self) -> CH_PRIOR_R

Bits 49:51 - Channel priority (7 is highest, 0 is lowest)

pub fn lock_ch(&self) -> LOCK_CH_R

Bit 52 - Channel lock bit

pub fn lock_ch_l(&self) -> LOCK_CH_L_R

Bits 53:54 - Channel lock level

pub fn src_osr_lmt(&self) -> SRC_OSR_LMT_R

Bits 55:58 - Source outstanding request limit

pub fn dst_osr_lmt(&self) -> DST_OSR_LMT_R

Bits 59:62 - Destination outstanding request limit