k210_pac::dmac::channel::cfg
Type Alias SRC_MULTBLK_TYPE_R

pub type SRC_MULTBLK_TYPE_R = R<u8, SRC_MULTBLK_TYPE_A>;

Reader of field src_multblk_type

Aliased Type

struct SRC_MULTBLK_TYPE_R { /* private fields */ }

Implementations

impl SRC_MULTBLK_TYPE_R

pub fn variant(&self) -> SRC_MULTBLK_TYPE_A

Get enumerated values variant

pub fn is_contiguous(&self) -> bool

Checks if the value of the field is CONTIGUOUS

pub fn is_reload(&self) -> bool

Checks if the value of the field is RELOAD

pub fn is_shadow_register(&self) -> bool

Checks if the value of the field is SHADOW_REGISTER

pub fn is_linked_list(&self) -> bool

Checks if the value of the field is LINKED_LIST