Type Definition k210_pac::dmac::channel::cfg::TT_FC_R

pub type TT_FC_R = R<u8, TT_FC_A>;

Reader of field tt_fc

Implementations

impl TT_FC_R

pub fn variant(&self) -> TT_FC_A

Get enumerated values variant

pub fn is_mem2mem_dma(&self) -> bool

Checks if the value of the field is MEM2MEM_DMA

pub fn is_mem2prf_dma(&self) -> bool

Checks if the value of the field is MEM2PRF_DMA

pub fn is_prf2mem_dma(&self) -> bool

Checks if the value of the field is PRF2MEM_DMA

pub fn is_prf2prf_dma(&self) -> bool

Checks if the value of the field is PRF2PRF_DMA

pub fn is_prf2mem_prf(&self) -> bool

Checks if the value of the field is PRF2MEM_PRF

pub fn is_prf2prf_srcprf(&self) -> bool

Checks if the value of the field is PRF2PRF_SRCPRF

pub fn is_mem2prf_prf(&self) -> bool

Checks if the value of the field is MEM2PRF_PRF

pub fn is_prf2prf_dstprf(&self) -> bool

Checks if the value of the field is PRF2PRF_DSTPRF