k210_pac::dvp::sccb_cfg
Type Alias BYTE_NUM_R

pub type BYTE_NUM_R = R<u8, BYTE_NUM_A>;

Reader of field byte_num

Aliased Type

struct BYTE_NUM_R { /* private fields */ }

Implementations

impl BYTE_NUM_R

pub fn variant(&self) -> Variant<u8, BYTE_NUM_A>

Get enumerated values variant

pub fn is_num2(&self) -> bool

Checks if the value of the field is NUM2

pub fn is_num3(&self) -> bool

Checks if the value of the field is NUM3

pub fn is_num4(&self) -> bool

Checks if the value of the field is NUM4