k210_pac
Module generic

Common register and bit access and modify traits

Structs

R

Register/field reader.

Reg

This structure provides volatile access to registers.

W

Register writer.

Enums

Variant

Used if enumerated values cover not the whole range.

Traits

Readable

Trait implemented by readable registers to enable the read method.

ResetValue

Reset value of the register.

Writable

Trait implemented by writeable registers.