k210_pac::generic
Trait ResetValue

pub trait ResetValue {
    type Type;

    // Required method
    fn reset_value() -> Self::Type;
}

Reset value of the register.

This value is the initial value for the write method. It can also be directly written to the register by using the reset method.

Required Associated Types

type Type

Raw register type (u8, u16, u32, …).

Required Methods

fn reset_value() -> Self::Type

Reset value of the register.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.