k210_pac::i2s0::ccr
Type Alias R

pub type R = R<u32, CCR>;

Reader of register ccr

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn clk_gate(&self) -> CLK_GATE_R

Bits 0:2 - Gating of sclk

pub fn clk_word_size(&self) -> CLK_WORD_SIZE_R

Bits 3:4 - The number of sclk cycles for which the word select line stayd in the left aligned or right aligned mode

pub fn align_mode(&self) -> ALIGN_MODE_R

Bits 5:7 - Alignment mode setting

pub fn dma_tx_en(&self) -> DMA_TX_EN_R

Bit 8 - DMA transmit enable control

pub fn dma_rx_en(&self) -> DMA_RX_EN_R

Bit 9 - DMA receive enable control

pub fn dma_divide_16(&self) -> DMA_DIVIDE_16_R

Bit 10 - Split 32bit data to two 16 bit data and filled in left and right channel. Used with dma_tx_en or dma_rx_en

pub fn sign_expand_en(&self) -> SIGN_EXPAND_EN_R

Bit 11 - SIGN_EXPAND_EN