Type Definition k210_pac::i2s0::channel::rcr::WLEN_R

pub type WLEN_R = R<u8, WLEN_A>;

Reader of field wlen

Implementations

impl WLEN_R

pub fn variant(&self) -> Variant<u8, WLEN_A>

Get enumerated values variant

pub fn is_ignore(&self) -> bool

Checks if the value of the field is IGNORE

pub fn is_resolution12(&self) -> bool

Checks if the value of the field is RESOLUTION12

pub fn is_resolution16(&self) -> bool

Checks if the value of the field is RESOLUTION16

pub fn is_resolution20(&self) -> bool

Checks if the value of the field is RESOLUTION20

pub fn is_resolution24(&self) -> bool

Checks if the value of the field is RESOLUTION24

pub fn is_resolution32(&self) -> bool

Checks if the value of the field is RESOLUTION32