k210_pac::i2s0::channel::rfcr
Type Alias RXCHDT_R

pub type RXCHDT_R = R<u8, RXCHDT_A>;

Reader of field rxchdt

Aliased Type

struct RXCHDT_R { /* private fields */ }

Implementations

impl RXCHDT_R

pub fn variant(&self) -> RXCHDT_A

Get enumerated values variant

pub fn is_level1(&self) -> bool

Checks if the value of the field is LEVEL1

pub fn is_level2(&self) -> bool

Checks if the value of the field is LEVEL2

pub fn is_level3(&self) -> bool

Checks if the value of the field is LEVEL3

pub fn is_level4(&self) -> bool

Checks if the value of the field is LEVEL4

pub fn is_level5(&self) -> bool

Checks if the value of the field is LEVEL5

pub fn is_level6(&self) -> bool

Checks if the value of the field is LEVEL6

pub fn is_level7(&self) -> bool

Checks if the value of the field is LEVEL7

pub fn is_level8(&self) -> bool

Checks if the value of the field is LEVEL8

pub fn is_level9(&self) -> bool

Checks if the value of the field is LEVEL9

pub fn is_level10(&self) -> bool

Checks if the value of the field is LEVEL10

pub fn is_level11(&self) -> bool

Checks if the value of the field is LEVEL11

pub fn is_level12(&self) -> bool

Checks if the value of the field is LEVEL12

pub fn is_level13(&self) -> bool

Checks if the value of the field is LEVEL13

pub fn is_level14(&self) -> bool

Checks if the value of the field is LEVEL14

pub fn is_level15(&self) -> bool

Checks if the value of the field is LEVEL15

pub fn is_level16(&self) -> bool

Checks if the value of the field is LEVEL16