k210_pac::plic::targets::threshold
Type Alias PRIORITY_R

pub type PRIORITY_R = R<u8, PRIORITY_A>;

Reader of field priority

Aliased Type

struct PRIORITY_R { /* private fields */ }

Implementations

impl PRIORITY_R

pub fn variant(&self) -> PRIORITY_A

Get enumerated values variant

pub fn is_never(&self) -> bool

Checks if the value of the field is NEVER

pub fn is_p1(&self) -> bool

Checks if the value of the field is P1

pub fn is_p2(&self) -> bool

Checks if the value of the field is P2

pub fn is_p3(&self) -> bool

Checks if the value of the field is P3

pub fn is_p4(&self) -> bool

Checks if the value of the field is P4

pub fn is_p5(&self) -> bool

Checks if the value of the field is P5

pub fn is_p6(&self) -> bool

Checks if the value of the field is P6

pub fn is_p7(&self) -> bool

Checks if the value of the field is P7