k210_pac::timer0::channel::control
Type Alias MODE_R

pub type MODE_R = R<bool, MODE_A>;

Reader of field mode

Aliased Type

struct MODE_R { /* private fields */ }

Implementations

impl MODE_R

pub fn variant(&self) -> MODE_A

Get enumerated values variant

pub fn is_free(&self) -> bool

Checks if the value of the field is FREE

pub fn is_user(&self) -> bool

Checks if the value of the field is USER