Type Definition k210_pac::dmac::channel::cfg::W

pub type W = W<u64, CFG>;

Writer for register cfg

Implementations

impl W

pub fn src_multblk_type(&mut self) -> SRC_MULTBLK_TYPE_W<'_>

Bits 0:1 - Source multi-block transfer type

pub fn dst_multblk_type(&mut self) -> DST_MULTBLK_TYPE_W<'_>

Bits 2:3 - Destination multi-block transfer type

pub fn tt_fc(&mut self) -> TT_FC_W<'_>

Bits 32:34 - Transfer type and flow control

pub fn hs_sel_src(&mut self) -> HS_SEL_SRC_W<'_>

Bit 35 - Source software or hardware handshaking select

pub fn hs_sel_dst(&mut self) -> HS_SEL_DST_W<'_>

Bit 36 - Destination software or hardware handshaking select

pub fn src_hwhs_pol(&mut self) -> SRC_HWHS_POL_W<'_>

Bit 37 - Source hardware handshaking interface polarity

pub fn dst_hwhs_pol(&mut self) -> DST_HWHS_POL_W<'_>

Bit 38 - Destination hardware handshaking interface polarity

pub fn src_per(&mut self) -> SRC_PER_W<'_>

Bits 39:42 - Assign a hardware handshaking interface to source of channel

pub fn dst_per(&mut self) -> DST_PER_W<'_>

Bits 44:47 - Assign a hardware handshaking interface to destination of channel

pub fn ch_prior(&mut self) -> CH_PRIOR_W<'_>

Bits 49:51 - Channel priority (7 is highest, 0 is lowest)

pub fn lock_ch(&mut self) -> LOCK_CH_W<'_>

Bit 52 - Channel lock bit

pub fn lock_ch_l(&mut self) -> LOCK_CH_L_W<'_>

Bits 53:54 - Channel lock level

pub fn src_osr_lmt(&mut self) -> SRC_OSR_LMT_W<'_>

Bits 55:58 - Source outstanding request limit

pub fn dst_osr_lmt(&mut self) -> DST_OSR_LMT_W<'_>

Bits 59:62 - Destination outstanding request limit