Type Definition k210_pac::dmac::channel::ctl::SRC_MSIZE_R

pub type SRC_MSIZE_R = R<u8, SRC_MSIZE_A>;

Reader of field src_msize

Implementations

impl SRC_MSIZE_R

pub fn variant(&self) -> Variant<u8, SRC_MSIZE_A>

Get enumerated values variant

pub fn is_length_1(&self) -> bool

Checks if the value of the field is LENGTH_1

pub fn is_length_4(&self) -> bool

Checks if the value of the field is LENGTH_4

pub fn is_length_8(&self) -> bool

Checks if the value of the field is LENGTH_8

pub fn is_length_16(&self) -> bool

Checks if the value of the field is LENGTH_16

pub fn is_length_32(&self) -> bool

Checks if the value of the field is LENGTH_32

pub fn is_length_64(&self) -> bool

Checks if the value of the field is LENGTH_64

pub fn is_length_128(&self) -> bool

Checks if the value of the field is LENGTH_128

pub fn is_length_256(&self) -> bool

Checks if the value of the field is LENGTH_256

pub fn is_length_512(&self) -> bool

Checks if the value of the field is LENGTH_512

pub fn is_length_1024(&self) -> bool

Checks if the value of the field is LENGTH_1024