Trait k210_pac::generic::ResetValue

pub trait ResetValue {
    type Type;
    fn reset_value() -> Self::Type;
}

Reset value of the register.

This value is the initial value for the write method. It can also be directly written to the register by using the reset method.

Associated Types

type Type

Raw register type (u8, u16, u32, …).

Required methods

fn reset_value() -> Self::Type

Reset value of the register.