Type Definition k210_pac::i2s0::ccr::W

pub type W = W<u32, CCR>;

Writer for register ccr

Implementations

impl W

pub fn clk_gate(&mut self) -> CLK_GATE_W<'_>

Bits 0:2 - Gating of sclk

pub fn clk_word_size(&mut self) -> CLK_WORD_SIZE_W<'_>

Bits 3:4 - The number of sclk cycles for which the word select line stayd in the left aligned or right aligned mode

pub fn align_mode(&mut self) -> ALIGN_MODE_W<'_>

Bits 5:7 - Alignment mode setting

pub fn dma_tx_en(&mut self) -> DMA_TX_EN_W<'_>

Bit 8 - DMA transmit enable control

pub fn dma_rx_en(&mut self) -> DMA_RX_EN_W<'_>

Bit 9 - DMA receive enable control

pub fn dma_divide_16(&mut self) -> DMA_DIVIDE_16_W<'_>

Bit 10 - Split 32bit data to two 16 bit data and filled in left and right channel. Used with dma_tx_en or dma_rx_en

pub fn sign_expand_en(&mut self) -> SIGN_EXPAND_EN_W<'_>

Bit 11 - SIGN_EXPAND_EN