Type Definition k210_pac::sysctl::dma_sel0::DMA_SEL0_R

pub type DMA_SEL0_R = R<u8, DMA_SEL0_A>;

Reader of field dma_sel0

Implementations

impl DMA_SEL0_R

pub fn variant(&self) -> Variant<u8, DMA_SEL0_A>

Get enumerated values variant

pub fn is_ssi0_rx_req(&self) -> bool

Checks if the value of the field is SSI0_RX_REQ

pub fn is_ssi0_tx_req(&self) -> bool

Checks if the value of the field is SSI0_TX_REQ

pub fn is_ssi1_rx_req(&self) -> bool

Checks if the value of the field is SSI1_RX_REQ

pub fn is_ssi1_tx_req(&self) -> bool

Checks if the value of the field is SSI1_TX_REQ

pub fn is_ssi2_rx_req(&self) -> bool

Checks if the value of the field is SSI2_RX_REQ

pub fn is_ssi2_tx_req(&self) -> bool

Checks if the value of the field is SSI2_TX_REQ

pub fn is_ssi3_rx_req(&self) -> bool

Checks if the value of the field is SSI3_RX_REQ

pub fn is_ssi3_tx_req(&self) -> bool

Checks if the value of the field is SSI3_TX_REQ

pub fn is_i2c0_rx_req(&self) -> bool

Checks if the value of the field is I2C0_RX_REQ

pub fn is_i2c0_tx_req(&self) -> bool

Checks if the value of the field is I2C0_TX_REQ

pub fn is_i2c1_rx_req(&self) -> bool

Checks if the value of the field is I2C1_RX_REQ

pub fn is_i2c1_tx_req(&self) -> bool

Checks if the value of the field is I2C1_TX_REQ

pub fn is_i2c2_rx_req(&self) -> bool

Checks if the value of the field is I2C2_RX_REQ

pub fn is_i2c2_tx_req(&self) -> bool

Checks if the value of the field is I2C2_TX_REQ

pub fn is_uart1_rx_req(&self) -> bool

Checks if the value of the field is UART1_RX_REQ

pub fn is_uart1_tx_req(&self) -> bool

Checks if the value of the field is UART1_TX_REQ

pub fn is_uart2_rx_req(&self) -> bool

Checks if the value of the field is UART2_RX_REQ

pub fn is_uart2_tx_req(&self) -> bool

Checks if the value of the field is UART2_TX_REQ

pub fn is_uart3_rx_req(&self) -> bool

Checks if the value of the field is UART3_RX_REQ

pub fn is_uart3_tx_req(&self) -> bool

Checks if the value of the field is UART3_TX_REQ

pub fn is_aes_req(&self) -> bool

Checks if the value of the field is AES_REQ

pub fn is_sha_rx_req(&self) -> bool

Checks if the value of the field is SHA_RX_REQ

pub fn is_ai_rx_req(&self) -> bool

Checks if the value of the field is AI_RX_REQ

pub fn is_fft_rx_req(&self) -> bool

Checks if the value of the field is FFT_RX_REQ

pub fn is_fft_tx_req(&self) -> bool

Checks if the value of the field is FFT_TX_REQ

pub fn is_i2s0_tx_req(&self) -> bool

Checks if the value of the field is I2S0_TX_REQ

pub fn is_i2s0_rx_req(&self) -> bool

Checks if the value of the field is I2S0_RX_REQ

pub fn is_i2s1_tx_req(&self) -> bool

Checks if the value of the field is I2S1_TX_REQ

pub fn is_i2s1_rx_req(&self) -> bool

Checks if the value of the field is I2S1_RX_REQ

pub fn is_i2s2_tx_req(&self) -> bool

Checks if the value of the field is I2S2_TX_REQ

pub fn is_i2s2_rx_req(&self) -> bool

Checks if the value of the field is I2S2_RX_REQ

pub fn is_i2s0_bf_dir_req(&self) -> bool

Checks if the value of the field is I2S0_BF_DIR_REQ

pub fn is_i2s0_bf_voice_req(&self) -> bool

Checks if the value of the field is I2S0_BF_VOICE_REQ